fix(Debuffs): Added new script Debuffmanager.cs and Modified BallProperties[2nd iteration].
testing with the new Debuff manager which would effect the player if he picks it up with various different buffs. the current implemented one doubles the size of the user for 5 seconds. which should make it harder for him if we place it in small platforms. this is only the first iteration. we have to set a template for it.
Modified ball properties. we do not need to put it on an object any more. we pass a gameObject to it in order to do various operations on it. thats how i did the double size. i got the object i.e the player game object when he touched the onTrigger of the debuff. passed it to the ballproperties. saved all the different variables. i.e scale, location, mass. and then performed the double size operation by using the new modifyscale() function. also added 3 new variables. scaleX scaleY and scaleZ. which effect the size of the object., now the ball properties has Mass, drag, angular drag, scale x y z and transform x y z. we can do many different operations with these values.