fix(Controls, Debuffs): Added new script BasicControls V3[ third iteration] and modified Ball properties. and Debuff manager.
Basic Controls V3: this is a new movement script. after watching monkey ball. and testing, the acceleration from slow movement is not as good, so fixed speed is better. this new script has fixed speed that can be changed currently would set to 15f to 20f. added functions that effect the speed which are called when the user touches the debuffs. will be discussed in the debuff section.

Basic Controls V2: this is obselete mostly now. will be deleted later. just kept it for reference.

BallProperties : this script is becoming more major and important. alot of changes has been made to it. instead of passing a game object to it. we interact with it directly by calling the gameobject that uses it. it is the main script that allows any other script to interact with the ball. so no other script would talk to the ball directly. they would have to pass a call from this script and it would manage it. examples will be given next in the debuff section.

DebuffManager : this script has two more debuffs added to it. as well as a new method to know which type of debuff it is. the current method is getting the debuff name from the object it is attached to  by the function DebuffChecker(), two more debuffs have been added. the first one is reverse control debuff. which reverses all the controls i.e  left becomes right and so on... the second one is another punishing debuff. which is a stunner (slow / half speed). if the user takes it it stuns him for 2 seconds. the other debuffs last for 5 seconds currently. this script DOES NOT interact with the ball directly .as explained. it goes through the ballproperties script that manages it.

two more prefabs added. Half speed and double size. u can test it by just dragging a script on any prefab. make sure that the on trigger is ticked. and it has Rigid body. untick the gravity so it wont fall.