fix(Debuffs): Added new Debuff "Speed Multipler"
multiplies the current speed by a set multipler value. 2 for double 3 for triple. it gets the value from the name of the prefab. currently the uploaded prefab is called Speed Multipler _2-5, which means 2 times the speed for 5 seconds. if we make it 10-20. it would be 10 times the current speed for 20 seconds. after the time is finished. the object returns to its original speed.