fix(CAMERA) added new script CameraScript2.cs
the new script follows the ball will have 2 variables. the distance and how high it should be as well as the sensitivity of the rotation. the mouse will be used to rotate. currently rotation is available on 270 degrees. will need to adjust it to 360 degrees capability later and some other improvements.