fix(Camera) added new script camerascript3.cs
now we can delete the first three scripts if we want to. this script is the most optimized so far. it should be able to scroll around the ball 360 degrees freely. we can control the speed of scrolling by phisensetivity varable from within the editor. and the distance of the camera with the radius variable also within the editor. this is the first iteration but this mainly will be the algorithm we go with. the next iteration will be doing the Y axis scrolling if we agree that we need to. just drag and drop this instead of the one you are using on the camera.